Pull out wallet row
